refactor(thoth): fold dom-info in as an internal module
Moves dom-info's source + tests into packages/thoth/dom-info/; Analyzer imports it relatively; thoth absorbs dom-info's 7 deps and drops the @azothjs/dom-info workspace dep. dom-info package removed.

Crucially: updates vite.config.js validationTests paths (packages/dom-info → packages/thoth/dom-info) so the browser-validation suites (dom-props/events/svg) stay EXCLUDED from the default/CI run — run on demand via 'pnpm test:validate', as before the move. resolve.test.js stays in the default run. Also drops the now-stale jsonic exclude. Lock regenerated with the pinned pnpm 11.5.2.

Verified: default run collects only dom-info/resolve.test.js; VALIDATE=true collects the 3 validation suites.
