+++ packages/thoth/dom-info/index.js
