+++ packages/thoth/dom-info/svg.test.js
