+++ packages/thoth/package.json
@@ -31,12 +31,18 @@
         "test": "vitest"
     },
     "dependencies": {
-        "@azothjs/dom-info": "workspace:*",
         "acorn": "^8.11.3",
         "acorn-jsx": "^5.3.2",
+        "aria-attributes": "^2.0.1",
         "astring": "^1.8.6",
+        "html-element-attributes": "^3.3.0",
+        "html-tag-names": "^2.1.0",
         "html-void-elements": "^3.0.0",
         "is-valid-es-identifier": "^2.0.0",
-        "source-map": "^0.7.4"
+        "mathml-tag-names": "^4.0.0",
+        "property-information": "^7.2.0",
+        "source-map": "^0.7.4",
+        "svg-element-attributes": "^2.1.0",
+        "svg-tag-names": "^3.0.1"
     }
 }
